Исправлено использование дефайна CHECK_HR. Добавлено значение по умолчанию для CVideoDecDXVAAllocator::m_nSurfaceArrayCount. И др.
git-svn-id: svn://mpc-next.ru/mpc-be/trunk@4275 d191f3aa-8897-45be-b9a4-56e5fb472b67

+++ src/filters/transform/VSFilter/Copy.cpp
@@ -217,7 +217,7 @@ void CDirectVobSubFilter::PrintMessages(BYTE* pOut)
 			tmp.Format(_T("queue stats: %I64d - %I64d [ms]\n"), rtStart/10000, rtStop/10000);
 			msg += tmp;
 
-			for (ptrdiff_t i = 0; i < nSubPics; i++) {
+			for (int i = 0; i < nSubPics; i++) {
 				m_pSubPicQueue->GetStats(i, rtStart, rtStop);
 				tmp.Format(_T("%d: %I64d - %I64d [ms]\n"), i, rtStart/10000, rtStop/10000);
 				msg += tmp;
